adding findbyTipoDocumentoAndNroDocumento

+++ src/main/java/co/edu/uco/uconotas/repository/TipoDocumentoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoDocumentoRepository extends JpaRepository<TipoDocumentoIdentidadEntity,Long> {
+public interface TipoDocumentoRepository extends JpaRepository<TipoDocumentoIdentidadEntity,Integer> {
 
 }
